Remove test_forget_call_super.py
